refactor(PostStatus): entity에서 dto로 패키지 이동
