+++ src/main/java/econo/buddybridge/comment/repository/CommentRepositoryImpl.java
@@ -13,8 +13,8 @@ import econo.buddybridge.comment.dto.QAuthorDto;
 import econo.buddybridge.comment.dto.QCommentResDto;
 import econo.buddybridge.comment.dto.QMyPageCommentResDto;
 import econo.buddybridge.matching.entity.Matching;
+import econo.buddybridge.post.dto.PostStatus;
 import econo.buddybridge.post.entity.Post;
-import econo.buddybridge.post.entity.PostStatus;
 import econo.buddybridge.post.entity.PostType;
 import econo.buddybridge.post.repository.PostRepositoryImpl;
 import lombok.RequiredArgsConstructor;
