+++ src/main/java/econo/buddybridge/matching/repository/MatchingRepositoryCustomImpl.java
@@ -16,8 +16,8 @@ import econo.buddybridge.member.entity.Member;
 import econo.buddybridge.member.entity.MemberRole;
 import econo.buddybridge.post.dto.CompletedVolunteerPostDto;
 import econo.buddybridge.post.dto.CompletedVolunteerPostPage;
+import econo.buddybridge.post.dto.PostStatus;
 import econo.buddybridge.post.dto.ScheduleDetailResDto;
-import econo.buddybridge.post.entity.PostStatus;
 import econo.buddybridge.post.entity.QPost;
 import econo.buddybridge.post.exception.PostInvalidSortValueException;
 import lombok.RequiredArgsConstructor;
