+++ src/main/java/econo/buddybridge/post/dto/CompletedVolunteerPostDto.java
@@ -4,7 +4,6 @@ import econo.buddybridge.matching.entity.MatchingStatus;
 import econo.buddybridge.member.entity.DisabilityType;
 import econo.buddybridge.post.entity.AssistanceType;
 import econo.buddybridge.post.entity.District;
-import econo.buddybridge.post.entity.PostStatus;
 import econo.buddybridge.post.entity.PostType;
 import lombok.Builder;
