+++ src/main/java/econo/buddybridge/post/dto/PostListItemDto.java
@@ -2,7 +2,6 @@ package econo.buddybridge.post.dto;
 
 import econo.buddybridge.member.entity.DisabilityType;
 import econo.buddybridge.post.entity.District;
-import econo.buddybridge.post.entity.PostStatus;
 import econo.buddybridge.post.entity.PostType;
 import lombok.Builder;
 
@@ -18,5 +17,5 @@ public record PostListItemDto(
         ScheduleListResDto schedule,
         Boolean isLiked
 ) {
-    
+
 }
