+++ src/main/java/econo/buddybridge/post/service/PostService.java
@@ -10,11 +10,11 @@ import econo.buddybridge.post.dto.PostCustomPage;
 import econo.buddybridge.post.dto.PostDetailDto;
 import econo.buddybridge.post.dto.PostEnumResDto;
 import econo.buddybridge.post.dto.PostReqDto;
+import econo.buddybridge.post.dto.PostStatus;
 import econo.buddybridge.post.dto.PostUpdateReqDto;
 import econo.buddybridge.post.entity.AssistanceType;
 import econo.buddybridge.post.entity.District;
 import econo.buddybridge.post.entity.Post;
-import econo.buddybridge.post.entity.PostStatus;
 import econo.buddybridge.post.entity.PostType;
 import econo.buddybridge.post.event.PostDeleteEvent;
 import econo.buddybridge.post.exception.PostDeleteNotAllowedException;
